requirements.txt
discord.py
python-dotenv
tinydb
flask
gspread
oauth2client